requirements.txt
# === Bot Dependencies (Legacy) ===
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
APScheduler>=3.10.0

# === FastAPI + PostgreSQL (New Architecture) ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
bcrypt>=4.1.0
PyJWT>=2.8.0

# File handling
python-multipart>=0.0.6
Pillow>=10.0.0
aiofiles>=23.0.0

# WebSocket
websockets>=12.0
